Laga et main program som me kan kjøre alle kombinasjoner av structures og solvers
